Fix issues with tty on vagrant Using local docker registry Using eloycoto/cilium base_box Clean k8s_install Add cilium compile script

+++ test/provision/k8s_install.sh
@@ -9,8 +9,10 @@ IP=$2
 K8S_VERSION=$3
 DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 
-apt-get update
-apt-get install -y curl jq apt-transport-https htop bmon
+if [[ -f  "/etc/provision_finished" ]]; then
+    /tmp/provision/compile.sh
+    exit 0
+fi
 
 cat <<EOF > /etc/hosts
 127.0.0.1       localhost
@@ -20,48 +22,26 @@ ff02::2 ip6-allrouters
 $IP     $NODE
 EOF
 
-
 cat <<EOF > /etc/apt/sources.list.d/kubernetes.list
 deb http://apt.kubernetes.io/ kubernetes-xenial main
 EOF
 
 curl https://packages.cloud.google.com/apt/doc/apt-key.gpg | apt-key add -
 
-curl -sSL https://get.docker.com/ | sh
-systemctl start docker
-
+apt-get update
 apt-get install --allow-downgrades -y \
+    llvm \
+    kubernetes-cni \
     kubelet="${K8S_VERSION}*" \
     kubeadm="${K8S_VERSION}*" \
-    kubectl="${K8S_VERSION}*" \
-    kubernetes-cni htop bmon
+    kubectl="${K8S_VERSION}*"
+
 
 sudo mkdir -p ${CILIUM_CONFIG_DIR}
 
-function install_etcd(){
-    wget -nv https://github.com/coreos/etcd/releases/download/${ETCD_VERSION}/etcd-${ETCD_VERSION}-linux-amd64.tar.gz
-    tar -xvf etcd-${ETCD_VERSION}-linux-amd64.tar.gz
-    sudo mv etcd-${ETCD_VERSION}-linux-amd64/etcd* /usr/bin/
-
-    sudo tee /etc/systemd/system/etcd.service <<EOF
-[Unit]
-Description=etcd
-Documentation=https://github.com/coreos
-
-[Service]
-ExecStart=/usr/bin/etcd --name=cilium --data-dir=/var/etcd/cilium --advertise-client-urls=http://192.168.36.11:9732 --listen-client-urls=http://0.0.0.0:9732 --listen-peer-urls=http://0.0.0.0:9733
-Restart=on-failure
-RestartSec=5
-
-[Install]
-WantedBy=multi-user.target
-EOF
-    sudo systemctl enable etcd
-    sudo systemctl start etcd
-}
-
 sudo mount bpffs /sys/fs/bpf -t bpf
 
+#check hostname to know if is kubernetes or runtime test
 if [[ "${HOST}" == "k8s1" ]]; then
     # FIXME: IP needs to be dynamic
     kubeadm init --token=$TOKEN --apiserver-advertise-address="192.168.36.11" --pod-network-cidr=10.10.0.0/16
@@ -77,10 +57,12 @@ if [[ "${HOST}" == "k8s1" ]]; then
     sudo cp /etc/kubernetes/admin.conf ${CILIUM_CONFIG_DIR}/kubeconfig
     kubectl taint nodes --all node-role.kubernetes.io/master-
 
-    install_etcd
+    sudo systemctl start etcd
+    /tmp/provision/compile.sh
 else
     kubeadm join --token=$TOKEN 192.168.36.11:6443
     cp /etc/kubernetes/kubelet.conf ${CILIUM_CONFIG_DIR}/kubeconfig
+    sudo systemctl stop etcd
 fi
 
-# install_etcd
+sudo touch /etc/provision_finished
